refactor(epics): share isSafeImageUrl for banner and DHO layout
CodeRabbit: dedupe URL safety checks with compact-space-banner.
