+++ packages/epics/src/spaces/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './hooks';
 export * from './utils/transparency-access';
+export { isSafeImageUrl } from './utils/safe-image-url';
